OK went a bit mad - far too much for a single commit.
1) Broke up the web endpoints into separate modules
2) Added new initial stack manager
3) updated breadcrumb bar to be full width
4) New tables for stack and project->technology link table
5) Changed /radar endpoints to be /category/xxx/technologies which is where they should have been
6) Fixed some of the links to ensure that they are URI encoded
