+++ routes/api/api-routes.js
@@ -2,15 +2,15 @@
  *  These are the routes implement the REST API services
  */
 
-var cache = require('../dao/cache.js');
-var users = require('../dao/users');
-var technology = require('../dao/technology.js');
-var category = require('../dao/category.js');
-var comments = require('../dao/comments.js');
-var projects = require('../dao/projects.js');
-var votes = require('../dao/vote.js');
-var status = require('../dao/status.js');
-var stacks = require('../dao/stacks.js');
+var cache = require('../../dao/cache.js');
+var users = require('../../dao/users');
+var technology = require('../../dao/technology.js');
+var category = require('../../dao/category.js');
+var comments = require('../../dao/comments.js');
+var projects = require('../../dao/projects.js');
+var votes = require('../../dao/vote.js');
+var status = require('../../dao/status.js');
+var stacks = require('../../dao/stacks.js');
 
 
 var bodyParser = require('body-parser');
@@ -18,7 +18,7 @@ var jsonParser = bodyParser.json();
 
 var passport = require('passport');
 
-var security = require('../utils/security.js');
+var security = require('../../utils/security.js');
 
 var sanitizer = require('sanitize-html');
