+++ views/pages/admin/addProject.ejs
@@ -24,7 +24,12 @@
 </div>
 
 <div class="container bs-docs-container">
-
+    <div class="col-md-12">
+        <ol class="breadcrumb" style="margin-top: 10px;">
+            <li><a href="/">Home</a></li>
+            <li class="active">Add Project</li>
+        </ol>
+    </div>
     <div class="row">
         <div class="col-md-6" style="padding-left: 15px; padding-top: 40px;">
             <form action="javascript:void(0);" method="POST">
