+++ views/pages/admin/addUser.ejs
@@ -25,7 +25,12 @@
 </div>
 
 <div class="container bs-docs-container">
-
+    <div class="col-md-12">
+        <ol class="breadcrumb" style="margin-top: 10px;">
+            <li><a href="/">Home</a></li>
+            <li class="active">Add User</li>
+        </ol>
+    </div>
     <div class="row">
         <div class="col-md-10 col-md-offset-1" style="padding-left: 15px; padding-top: 40px;">
             <form action="javascript:void(0);" method="POST">
