+++ views/pages/admin/listCategories.ejs
@@ -32,7 +32,7 @@
 </div>
 
 <div class="container bs-docs-container">
-    <div class="col-md-10 col-md-offset-1">
+    <div class="col-md-12">
         <ol class="breadcrumb" style="margin-top: 10px;">
             <li><a href="/">Home</a></li>
             <li class="active">Manage categories</li>
