+++ views/pages/editTechnology.ejs
@@ -20,7 +20,7 @@
 <% include ../partials/technology-head.ejs %>
 
 <div class="container bs-docs-container">
-    <div class="col-md-10 col-md-offset-1">
+    <div class="col-md-12">
         <ol class="breadcrumb" style="margin-top: 10px;">
             <li><a href="/">Home</a></li>
             <li><a href="/radar/<%= technology.categoryname.toLowerCase() %>"><%= technology.categoryname %></a></li>
