+++ views/pages/search.ejs
@@ -44,7 +44,7 @@
 
 <div class="container bs-docs-container">
 
-    <div class="col-md-10 col-md-offset-1">
+    <div class="col-md-12">
         <ol class="breadcrumb" style="margin-top: 10px;">
             <li><a href="/">Home</a></li>
             <li class="active">Search</li>
