+++ views/pages/votehistory.ejs
@@ -26,7 +26,7 @@
 <% include ../partials/technology-head.ejs %>
 
 <div class="container bs-docs-container" style="padding-top: 10px;">
-    <div class="col-md-10 col-md-offset-1">
+    <div class="col-md-12">
         <ol class="breadcrumb" style="margin-bottom: 5px;">
             <li><a href="/">Home</a></li>
             <li><a href="/radar/<%=getBreadcrumb() %>"><%=technology.categoryname%></a></li>
