+++ views/partials/menu.ejs
@@ -18,7 +18,7 @@
                     <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Reports<span class="caret"></span></a>
                     <ul class="dropdown-menu">
                         <% categories.forEach( function( category ) {
-                            var url = encodeURI('/radar/' + category.name ).toLowerCase(); %>
+                            var url = encodeURI('/category/' + category.name + '/technologies').toLowerCase(); %>
                             <li><a href="<%=url%>"><%=category.name%></a></li>
                         <% }); %>
